% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{plot.ltc}
\alias{plot.ltc}
\title{Plot a Colour Palette}
\usage{
\method{plot}{ltc}(x, ...)
}
\arguments{
\item{x}{A vector of colours from one of the \code{ltc} palettes (palette object).}

\item{...}{Additional arguments (currently unused).}
}
\value{
A ggplot2 object showing the selected colours.
}
\description{
Visualizes a selected colour palette as a bar of colours.
}
\examples{
\dontrun{
paloma <- ltc("paloma")
plot.ltc(paloma)
}
}
