% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{plot.ltc}
\alias{plot.ltc}
\title{Plot a Colour Palette}
\usage{
\method{plot}{ltc}(x, ...)
}
\arguments{
\item{x}{An ltc palette object}

\item{...}{Additional arguments (currently unused).}
}
\value{
A ggplot2 object showing the selected colours.
}
\description{
Visualizes a selected colour palette as a bar of colours.
}
\examples{
\donttest{
# Create and plot a palette
pal <- ltc(paloma)
plot(pal)
}
}
